[Refactor] #58 - 공통 로직 처리를 위한 BaseCoordinator 구현
